fix(cli): rename to import_core
