+++ crates/cli/commands/src/lib.rs
@@ -16,7 +16,7 @@ pub mod dump_genesis;
 pub mod export_era;
 pub mod import;
 pub mod import_era;
-pub mod import_op;
+pub mod import_core;
 pub mod init_cmd;
 pub mod init_state;
 pub mod launcher;
